requirements.txt
django==1.11.5
djangorestframework==3.6.4
markdown==2.6.9
django-filter==1.0.4
pytest-django
